# apps/web/openapi/petclinic-api.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "Petclinic MSS301 — Aggregated API",
    "version": "1.0.0",
    "description": "Aggregated from: auth, customers, vets, visits, workflow, billing, products, reviews"
  },
  "servers": [
    {
      "url": "http://localhost:8180",
      "description": "API Gateway"
    }
  ],
...
  "paths": {
...
    "/api/v1/products": {
      "get": {
        "tags": ["Products"],
        "summary": "List sản phẩm — lọc q (code/name), type, active, lowStock",
        "operationId": "listProducts",
        "parameters": [
          {
            "name": "q",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "type",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "enum": ["MEDICATION", "VACCINE", "SERVICE", "SUPPLY", "MERCHANDISE"]
            }
          },
          {
            "name": "active",
            "in": "query",
            "required": false,
            "schema": {
              "type": "boolean"
            }
          },
          {
            "name": "lowStock",
            "in": "query",
            "required": false,
            "schema": {
              "type": "boolean"
            }
          },
          {
            "name": "pageable",
            "in": "query",
            "required": true,
            "schema": {
              "$ref": "#/components/schemas/Pageable"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PageProductResponse"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": ["Products"],
        "summary": "Tạo sản phẩm mới (ADMIN/INVENTORY_MANAGER)",
        "operationId": "createProduct",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateProductRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProductResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/products/{id}/restock": {
      "post": {
        "tags": ["Products"],
        "summary": "Nhập thêm tồn kho (ADMIN/INVENTORY_MANAGER)",
        "operationId": "restockProduct",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          },
          {
            "name": "Idempotency-Key",
            "in": "header",
            "required": false,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/StockAdjustRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProductResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/products/{id}/consume": {
      "post": {
        "tags": ["Products"],
        "summary": "Trừ tồn kho — cấp phát thuốc/vật tư (STAFF/ADMIN/VET, dùng nội bộ khi kê đơn)",
        "operationId": "consumeProduct",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          },
          {
            "name": "Idempotency-Key",
            "in": "header",
            "required": false,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/StockAdjustRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProductResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/products/stock/documents": {
      "post": {
        "tags": ["Inventory"],
        "summary": "Record an audited multi-line stock receipt or issue document",
        "operationId": "recordManualInventoryDocument",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ManualStockDocumentRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/InventoryOperationResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/products/stock/consume": {
      "post": {
        "tags": ["Inventory"],
        "summary": "Atomically consume stock for one or more products",
        "operationId": "consume",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/BatchStockConsumeRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/InventoryOperationResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/products/{id}": {
      "get": {
        "tags": ["Products"],
        "summary": "Chi tiết sản phẩm",
        "operationId": "getProduct",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProductResponse"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": ["Products"],
        "summary": "Ngừng kinh doanh sản phẩm, giữ lịch sử (ADMIN/INVENTORY_MANAGER)",
        "operationId": "deleteProduct",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "No Content"
          }
        }
      },
      "patch": {
        "tags": ["Products"],
        "summary": "Cập nhật sản phẩm (ADMIN/INVENTORY_MANAGER)",
        "operationId": "updateProduct",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateProductRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProductResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/products/stock/{productId}/movements": {
      "get": {
        "tags": ["Inventory"],
        "summary": "List append-only stock movements for a product",
        "operationId": "listMovements",
        "parameters": [
          {
            "name": "productId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          },
          {
            "name": "pageable",
            "in": "query",
            "required": true,
            "schema": {
              "$ref": "#/components/schemas/Pageable"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PageStockMovementResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/products/stock/movements": {
      "get": {
        "tags": ["Inventory"],
        "summary": "List append-only stock movements across all products",
        "operationId": "listInventoryMovements",
        "parameters": [
          {
            "name": "pageable",
            "in": "query",
            "required": true,
            "schema": {
              "$ref": "#/components/schemas/Pageable"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PageStockMovementResponse"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "Pageable": {
        "type": "object",
        "properties": {
          "page": {
            "type": "integer",
            "format": "int32",
            "minimum": 0
          },
          "size": {
            "type": "integer",
            "format": "int32",
            "minimum": 1
          },
          "sort": {
            "type": "array",
            "items": {
              "type": "string"
            }
          }
        }
      },
...
      "PageableObject": {
        "type": "object",
        "properties": {
          "offset": {
            "type": "integer",
            "format": "int64"
          },
          "unpaged": {
            "type": "boolean"
          },
          "sort": {
            "$ref": "#/components/schemas/SortObject"
          },
          "paged": {
            "type": "boolean"
          },
          "pageNumber": {
            "type": "integer",
            "format": "int32"
          },
          "pageSize": {
            "type": "integer",
            "format": "int32"
          }
        }
      },
      "SortObject": {
        "type": "object",
        "properties": {
          "empty": {
            "type": "boolean"
          },
          "unsorted": {
            "type": "boolean"
          },
          "sorted": {
            "type": "boolean"
          }
        }
      },
...
      "CreateProductRequest": {
        "type": "object",
        "properties": {
          "code": {
            "type": "string",
            "maxLength": 50,
            "minLength": 0,
            "pattern": "[A-Za-z0-9_]+"
          },
          "name": {
            "type": "string",
            "maxLength": 150,
            "minLength": 0
          },
          "category": {
            "type": "string",
            "maxLength": 80,
            "minLength": 0
          },
          "description": {
            "type": "string",
            "maxLength": 2000,
            "minLength": 0
          },
          "type": {
            "type": "string",
            "enum": ["MEDICATION", "VACCINE", "SERVICE", "SUPPLY", "MERCHANDISE"]
          },
          "unitPrice": {
            "type": "number",
            "minimum": 0
          },
          "unit": {
            "type": "string",
            "maxLength": 30,
            "minLength": 0
          },
          "stockQuantity": {
            "type": "integer",
            "format": "int32"
          },
          "reorderLevel": {
            "type": "integer",
            "format": "int32"
          }
        },
        "required": ["code", "name", "type", "unitPrice"]
      },
      "ProductResponse": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "code": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "category": {
            "type": "string"
          },
          "description": {
            "type": "string"
          },
          "type": {
            "type": "string",
            "enum": ["MEDICATION", "VACCINE", "SERVICE", "SUPPLY", "MERCHANDISE"]
          },
          "unitPrice": {
            "type": "number"
          },
          "unit": {
            "type": "string"
          },
          "stockQuantity": {
            "type": "integer",
            "format": "int32"
          },
          "reorderLevel": {
            "type": "integer",
            "format": "int32"
          },
          "stockTracked": {
            "type": "boolean"
          },
          "stockStatus": {
            "type": "string",
            "enum": ["AVAILABLE", "LOW", "OUT"]
          },
          "active": {
            "type": "boolean"
          }
        }
      },
      "StockAdjustRequest": {
        "type": "object",
        "properties": {
          "quantity": {
            "type": "integer",
            "format": "int32"
          }
        },
        "required": ["quantity"]
      },
      "Line": {
        "type": "object",
        "properties": {
          "productId": {
            "type": "integer",
            "format": "int64"
          },
          "quantity": {
            "type": "integer",
            "format": "int32"
          }
        },
        "required": ["productId", "quantity"]
      },
      "ManualStockDocumentRequest": {
        "type": "object",
        "properties": {
          "idempotencyKey": {
            "type": "string",
            "maxLength": 160,
            "minLength": 0
          },
          "direction": {
            "type": "string",
            "enum": ["IN", "OUT"]
          },
          "reason": {
            "type": "string",
            "maxLength": 255,
            "minLength": 0
          },
          "reference": {
            "type": "string",
            "maxLength": 120,
            "minLength": 0
          },
          "items": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Line"
            },
            "maxItems": 100,
            "minItems": 0
          }
        },
        "required": ["direction", "idempotencyKey", "items", "reason"]
      },
      "InventoryOperationResponse": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "idempotencyKey": {
            "type": "string"
          },
          "operationType": {
            "type": "string",
            "enum": ["INITIAL", "CONSUME", "RESTOCK", "ADJUSTMENT"]
          },
          "sourceType": {
            "type": "string"
          },
          "sourceId": {
            "type": "string"
          },
          "reason": {
            "type": "string"
          },
          "movements": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/StockMovementResponse"
            }
          }
        }
      },
      "StockMovementResponse": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "operationId": {
            "type": "integer",
            "format": "int64"
          },
          "idempotencyKey": {
            "type": "string"
          },
          "operationType": {
            "type": "string",
            "enum": ["INITIAL", "CONSUME", "RESTOCK", "ADJUSTMENT"]
          },
          "sourceType": {
            "type": "string"
          },
          "sourceId": {
            "type": "string"
          },
          "reason": {
            "type": "string"
          },
          "productId": {
            "type": "integer",
            "format": "int64"
          },
          "productCode": {
            "type": "string"
          },
          "quantityDelta": {
            "type": "integer",
            "format": "int32"
          },
          "quantityBefore": {
            "type": "integer",
            "format": "int32"
          },
          "quantityAfter": {
            "type": "integer",
            "format": "int32"
          },
          "createdBy": {
            "type": "string"
          },
          "createdDate": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
      "BatchStockConsumeRequest": {
        "type": "object",
        "properties": {
          "idempotencyKey": {
            "type": "string",
            "maxLength": 160,
            "minLength": 0
          },
          "sourceType": {
            "type": "string",
            "maxLength": 40,
            "minLength": 0
          },
          "sourceId": {
            "type": "string",
            "maxLength": 120,
            "minLength": 0
          },
          "reason": {
            "type": "string",
            "maxLength": 255,
            "minLength": 0
          },
          "items": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Line"
            },
            "maxItems": 100,
            "minItems": 0
          }
        },
        "required": ["idempotencyKey", "items", "sourceType"]
      },
      "UpdateProductRequest": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "maxLength": 150,
            "minLength": 0,
            "pattern": ".*\\S.*"
          },
          "category": {
            "type": "string",
            "maxLength": 80,
            "minLength": 0
          },
          "description": {
            "type": "string",
            "maxLength": 2000,
            "minLength": 0
          },
          "unitPrice": {
            "type": "number",
            "minimum": 0
          },
          "unit": {
            "type": "string",
            "maxLength": 30,
            "minLength": 0
          },
          "stockQuantity": {
            "type": "integer",
            "format": "int32"
          },
          "reorderLevel": {
            "type": "integer",
            "format": "int32"
          },
          "active": {
            "type": "boolean"
          }
        }
      },
      "PageProductResponse": {
        "type": "object",
        "properties": {
          "totalElements": {
            "type": "integer",
            "format": "int64"
          },
          "totalPages": {
            "type": "integer",
            "format": "int32"
          },
          "size": {
            "type": "integer",
            "format": "int32"
          },
          "content": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ProductResponse"
            }
          },
          "number": {
            "type": "integer",
            "format": "int32"
          },
          "first": {
            "type": "boolean"
          },
          "last": {
            "type": "boolean"
          },
          "numberOfElements": {
            "type": "integer",
            "format": "int32"
          },
          "pageable": {
            "$ref": "#/components/schemas/PageableObject"
          },
          "sort": {
            "$ref": "#/components/schemas/SortObject"
          },
          "empty": {
            "type": "boolean"
          }
        }
      },
      "PageStockMovementResponse": {
        "type": "object",
        "properties": {
          "totalElements": {
            "type": "integer",
            "format": "int64"
          },
          "totalPages": {
            "type": "integer",
            "format": "int32"
          },
          "size": {
            "type": "integer",
            "format": "int32"
          },
          "content": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/StockMovementResponse"
            }
          },
          "number": {
            "type": "integer",
            "format": "int32"
          },
          "first": {
            "type": "boolean"
          },
          "last": {
            "type": "boolean"
          },
          "numberOfElements": {
            "type": "integer",
            "format": "int32"
          },
          "pageable": {
            "$ref": "#/components/schemas/PageableObject"
          },
          "sort": {
            "$ref": "#/components/schemas/SortObject"
          },
          "empty": {
            "type": "boolean"
          }
        }
      },
...
    },
    "securitySchemes": {}
  }
}